+++ src/flscan312.cpp
@@ -60,6 +60,7 @@ int CLUSTERTYPE = 0;
 int     K = 15;         /* maximum length of connected areas */
 int		K2;
 int     N;              /* number of all areas */
+double CLUSTERRADIUS = DBL_MAX; /* maximum radius of connected areas */
 
 areaidx	*w;             /* w[] */
 areaidx	z[KLIM];        /* z[] */
@@ -422,6 +423,10 @@ void FlexibleScan0s(int zlen) {
     /* if already included or finished then skip. */
     if (masksw[caz[i]] != 0) /* -1:included, 1,2,3..:finished */
     continue;
+    
+    if (area[caz[i]].dist > CLUSTERRADIUS)
+      continue;
+    
     /* add an area to z[] */
     z[zlen] = r = caz[i];
     masksw[r] = -1;
@@ -485,6 +490,9 @@ void	CircularScan0s(int zlen) {
   };
   
   if (zlen == K2)
+    return;
+  
+  if (area[w[zlen]].dist > CLUSTERRADIUS)
     return;
   
   r = w[zlen];
@@ -553,6 +561,10 @@ void FlexibleScan1s(int zlen, int ss) {
     /* if already included or finished then skip. */
     if (masksw[caz[i]] != 0) /* -1:included, 1,2,3..:finished */
     continue;
+    
+    if (area[caz[i]].dist > CLUSTERRADIUS)
+      continue;
+    
     /* add an area to z[] */
     z[zlen] = r = caz[i];
     masksw[r] = -1;
@@ -627,6 +639,9 @@ void	CircularScan1s(int zlen, int ss) {
   };
   
   if (zlen == K2)
+    return;
+  
+  if (area[w[zlen]].dist > CLUSTERRADIUS)
     return;
   
   r = w[zlen];
@@ -742,6 +757,10 @@ void FlexibleScan0l(int zlen) {
     /* if already included or finished then skip. */
     if (masksw[caz[i]] != 0) /* -1:included, 1,2,3..:finished */
     continue;
+    
+    if (area[caz[i]].dist > CLUSTERRADIUS)
+      continue;
+    
     /* add an area to z[] */
     z[zlen] = r = caz[i];
     masksw[r] = -1;
@@ -799,6 +818,9 @@ void	CircularScan0l(int zlen) {
   };
   
   if (zlen == K2)
+    return;
+  
+  if (area[w[zlen]].dist > CLUSTERRADIUS)
     return;
   
   r = w[zlen];
@@ -859,6 +881,10 @@ void FlexibleScan1l(int zlen, int ss) {
     /* if already included or finished then skip. */
     if (masksw[caz[i]] != 0) /* -1:included, 1,2,3..:finished */
     continue;
+    
+    if (area[caz[i]].dist > CLUSTERRADIUS)
+      continue;
+    
     /* add an area to z[] */
     z[zlen] = r = caz[i];
     masksw[r] = -1;
@@ -926,6 +952,9 @@ void	CircularScan1l(int zlen, int ss) {
   };
   
   if (zlen == K2)
+    return;
+  
+  if (area[w[zlen]].dist > CLUSTERRADIUS)
     return;
   
   r = w[zlen];
@@ -1991,6 +2020,7 @@ List runFleXScan(const List &setting,
   }
 
   CLUSTERTYPE = setting["clustertype"];
+  CLUSTERRADIUS = setting["clusterradius"];
   
   switch(CLUSTERTYPE) {
     case HOT: